Refactored configureWebpack.js
