+++ packages/pwa-buildpack/lib/WebpackTools/configureWebpack.js
@@ -1,17 +1,11 @@
 const { promisify } = require('util');
 const stat = promisify(require('fs').stat);
 const path = require('path');
-const webpack = require('webpack');
 const pkgDir = require('pkg-dir');
-const TerserPlugin = require('terser-webpack-plugin');
-const WebpackAssetsManifest = require('webpack-assets-manifest');
 
 const loadEnvironment = require('../Utilities/loadEnvironment');
-const RootComponentsPlugin = require('./plugins/RootComponentsPlugin');
-const ServiceWorkerPlugin = require('./plugins/ServiceWorkerPlugin');
-const UpwardIncludePlugin = require('./plugins/UpwardIncludePlugin');
-const PWADevServer = require('./PWADevServer');
-const MagentoResolver = require('./MagentoResolver');
+const getClientConfig = require('../Utilities/getClientConfig');
+const getServiceWorkerConfig = require('../Utilities/getServiceWorkerConfig');
 
 /**
  * We need a root directory for the app in order to build all paths relative to
@@ -58,10 +52,6 @@ function getMode(cliEnv = {}, projectConfig) {
     return 'development';
 }
 
-function isDevServer() {
-    return process.argv.find(v => v.includes('webpack-dev-server'));
-}
-
 async function configureWebpack({ context, vendor = [], special = {}, env }) {
     await validateRoot(context);
 
@@ -90,231 +80,23 @@ async function configureWebpack({ context, vendor = [], special = {}, env }) {
 
     const mode = getMode(env, projectConfig);
 
-    const config = {
+    const configOptions = {
         mode,
-        context, // Node global for the running script's directory
-        entry: {
-            client: path.resolve(paths.src, 'index.js')
-        },
-        output: {
-            path: paths.output,
-            publicPath: '/',
-            filename: '[name].js',
-            strictModuleExceptionHandling: true,
-            chunkFilename: '[name]-[chunkhash].js'
-        },
-        module: {
-            rules: [
-                {
-                    test: /\.graphql$/,
-                    include: [paths.src, ...hasFlag('graphqlQueries')],
-                    use: [
-                        {
-                            loader: 'graphql-tag/loader'
-                        }
-                    ]
-                },
-                {
-                    test: /\.(mjs|js)$/,
-                    include: [paths.src, ...hasFlag('esModules')],
-                    sideEffects: false,
-                    use: [
-                        {
-                            loader: 'babel-loader',
-                            options: {
-                                envName: mode,
-                                rootMode: babelConfigPresent ? 'root' : 'upward'
-                            }
-                        }
-                    ]
-                },
-                {
-                    test: /\.css$/,
-                    oneOf: [
-                        {
-                            test: [paths.src, ...hasFlag('cssModules')],
-                            use: [
-                                'style-loader',
-                                {
-                                    loader: 'css-loader',
-                                    options: {
-                                        localIdentName:
-                                            '[name]-[local]-[hash:base64:3]',
-                                        modules: true
-                                    }
-                                }
-                            ]
-                        },
-                        {
-                            include: /node_modules/,
-                            use: [
-                                'style-loader',
-                                {
-                                    loader: 'css-loader',
-                                    options: {
-                                        modules: false
-                                    }
-                                }
-                            ]
-                        }
-                    ]
-                },
-                {
-                    test: /\.(jpg|svg)$/,
-                    use: [
-                        {
-                            loader: 'file-loader',
-                            options: {
-                                name: '[name]-[hash:base58:3].[ext]'
-                            }
-                        }
-                    ]
-                }
-            ]
-        },
-        resolve: await MagentoResolver.configure({
-            paths: {
-                root: context
-            }
-        }),
-        plugins: [
-            new RootComponentsPlugin({
-                rootComponentsDirs: [
-                    ...hasFlag('rootComponents'),
-                    context
-                ].reduce(
-                    (searchPaths, moduleDir) => [
-                        ...searchPaths,
-                        path.join(moduleDir, 'RootComponents'),
-                        path.join(moduleDir, 'src', 'RootComponents'),
-                        path.join(moduleDir, 'lib', 'RootComponents')
-                    ],
-                    []
-                ),
-                context
-            }),
-            new webpack.EnvironmentPlugin(projectConfig.env),
-            new ServiceWorkerPlugin({
-                mode,
-                paths,
-                injectManifest: true,
-                injectManifestConfig: {
-                    include: [/\.js$/],
-                    swSrc: './src/ServiceWorker/sw.js',
-                    swDest: './sw.js'
-                }
-            }),
-            new UpwardIncludePlugin({
-                upwardDirs: [...hasFlag('upward'), context]
-            }),
-            new WebpackAssetsManifest({
-                output: 'asset-manifest.json',
-                entrypoints: true,
-                publicPath: '/',
-                // Add explicit properties to the asset manifest for
-                // upward.yml to use when evaluating app shell templates.
-                transform(assets) {
-                    // All RootComponents go to prefetch, and all client scripts
-                    // go to load.
-                    assets.bundles = {
-                        load: assets.entrypoints.client.js,
-                        prefetch: []
-                    };
-                    Object.entries(assets).forEach(([name, value]) => {
-                        if (name.match(/^RootCmp.*\.js$/)) {
-                            const filenames = Array.isArray(value)
-                                ? value
-                                : [value];
-                            assets.bundles.prefetch.push(...filenames);
-                        }
-                        const ext = path.extname(name);
-                        const type = ext && ext.replace(/^\./, '');
-                        if (type) {
-                            if (!assets[type]) {
-                                assets[type] = {};
-                            }
-                            assets[type][path.basename(name, ext)] = value;
-                        }
-                    });
-                }
-            })
-        ],
-        devtool: 'source-map'
+        context,
+        babelConfigPresent,
+        paths,
+        hasFlag
     };
-    let vendorTest = '[\\/]node_modules[\\/]';
-    if (vendor.length > 0) {
-        vendorTest += `(${vendor.join('|')})[\\\/]`;
-    }
-    config.optimization = {
-        splitChunks: {
-            cacheGroups: {
-                vendor: {
-                    test: new RegExp(vendorTest),
-                    chunks: 'all'
-                }
-            }
-        }
-    };
-    if (mode === 'development') {
-        Object.assign(config.optimization, {
-            moduleIds: 'named',
-            nodeEnv: 'development',
-            minimize: false,
-            occurrenceOrder: true,
-            usedExports: true,
-            concatenateModules: true,
-            sideEffects: true
-        });
 
-        if (isDevServer()) {
-            // Using eval-source-map shows original source (non-transpiled) as
-            // well as comments.
-            // See https://webpack.js.org/configuration/devtool/
-            config.devtool = 'eval-source-map';
+    const serviceWorkerConfig = getServiceWorkerConfig(configOptions);
 
-            await PWADevServer.configure(
-                {
-                    graphqlPlayground: true,
-                    ...projectConfig.sections(
-                        'devServer',
-                        'imageService',
-                        'customOrigin'
-                    ),
-                    ...projectConfig.section('magento'),
-                    upwardPath: projectConfig.section('upwardJs').upwardPath
-                },
-                config
-            );
-        }
-    } else if (mode === 'production') {
-        config.performance = {
-            hints: 'warning'
-        };
-        config.devtool = false;
-        config.optimization.minimizer = [
-            new TerserPlugin({
-                parallel: true,
-                cache: true,
-                terserOptions: {
-                    ecma: 8,
-                    parse: {
-                        ecma: 8
-                    },
-                    compress: {
-                        drop_console: true
-                    },
-                    output: {
-                        ecma: 7,
-                        semicolons: false
-                    },
-                    keep_fnames: true
-                }
-            })
-        ];
-    } else {
-        throw Error(`Unsupported environment mode in webpack config: ${mode}`);
-    }
-    return config;
+    const clientConfig = await getClientConfig({
+        ...configOptions,
+        vendor,
+        projectConfig
+    });
+
+    return { clientConfig, serviceWorkerConfig };
 }
 
 module.exports = configureWebpack;
